changed 'int' to 'double'

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Intake.java
@@ -15,7 +15,7 @@ public class Intake extends Subsystem{
         this.wheels = wheels;
     }
 
-    public void runIntake(int power)
+    public void runIntake(double power)
     {
         wheels.setPower(power);
     }
